core: users: fixup system-user example
The newline in the `assertion ="""` line passes an unparseable assertion
to snapd, resulting in syntax error messages. Removing this newline
allows the assertion to be successfully acknowledged by snapd.

In order for this assertion to be acknowledged by snapd, it must
also have acknowledged the account and account-key assertions for the
authority-id and signing key of the assertion. snapd itself supports
the --chain flag to append these assertions to the output of the sign
command, so the example now accurately reflects what should be submitted
to snapd.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>